refactor: Split worker code and `StreamHandler` in to co-located files
