+++ runner/src/stream-handler/index.ts
@@ -1,0 +1,1 @@
+export { default } from './stream-handler';
